// Project/gamanza-challenge/src/api/types.ts
//product 
export interface Product {
  id: number;
  title: string;
  description: string;
  price: number;
  discountPercentage: number;
  rating: number;
  stock: number;
  brand: string;
  category: string;
  thumbnail: string;
  images: string[];
}

//  (paginated)
export interface ProductsResponse {
  products: Product[];
  total: number;
  skip: number;
  limit: number;
}